Clone Method
 The method Object.clone() (from java.lang.Object) is a protected
method, and it throws a CloneNotSupportedException if the class
does not implement the Cloneable interface.  � Here's what happens internally:
 super.clone() is called, which is Object.clone().  The Object.clone() method:
o Creates a new object in memory. o Copies all the primitive fields (int a) from d1 to the new
object d2. o For object references (if you had any), it copies the
reference, not the actual object (i.e., shallow copy). � What is Cloneable?
Cloneable is a marker interface in Java, meaning:  It has no methods.  It simply marks a class as being eligible for cloning via the
Object.clone() method.